+++ src/Streams.php
@@ -9,7 +9,16 @@ final class Streams implements Capabilities
     {
     }
 
+    /**
+     * @deprecated
+     * @see ::fromAmbientAuthority()
+     */
     public static function of(): self
+    {
+        return new self;
+    }
+
+    public static function fromAmbientAuthority(): self
     {
         return new self;
     }