+++ tests/StreamsTest.php
@@ -19,13 +19,13 @@ class StreamsTest extends TestCase
     {
         $this->assertInstanceOf(
             Capabilities::class,
-            Streams::of(),
+            Streams::fromAmbientAuthority(),
         );
     }
 
     public function testOpeningATemporaryStreamAlwaysReturnANewOne()
     {
-        $streams = Streams::of();
+        $streams = Streams::fromAmbientAuthority();
         $a = $streams->temporary()->new()->write(Str::of('a'))->match(
             static fn($a) => $a,
             static fn() => null,
@@ -56,7 +56,7 @@ class StreamsTest extends TestCase
 
     public function testOpenReadable()
     {
-        $self = Streams::of()
+        $self = Streams::fromAmbientAuthority()
             ->readable()
             ->open(Path::of(__FILE__))
             ->toString()
@@ -71,7 +71,7 @@ class StreamsTest extends TestCase
     public function testOpenWritable()
     {
         $path = Path::of(\tempnam(\sys_get_temp_dir(), 'streams'));
-        $streams = Streams::of();
+        $streams = Streams::fromAmbientAuthority();
         $streams
             ->writable()
             ->open($path)
@@ -94,13 +94,13 @@ class StreamsTest extends TestCase
     {
         $this->assertInstanceOf(
             Select::class,
-            Streams::of()
+            Streams::fromAmbientAuthority()
                 ->watch()
                 ->waitForever(),
         );
         $this->assertInstanceOf(
             Select::class,
-            Streams::of()
+            Streams::fromAmbientAuthority()
                 ->watch()
                 ->timeoutAfter(new ElapsedPeriod(1)),
         );